fixed delete candy endpoint
The issue lied in the dependencies that candy had( both wishlist and order depended on a candy id),  therefore if a candy is deleted all that the candy from wishlist and order should be deleted. To add to that, we could use cascade, but the issue again is that the repository operation wont be synced. Refer to this for more info: https://stackoverflow.com/questions/22688402/delete-not-working-with-jparepository